refactor(ivy): move `styling/instructions.ts` to `instructions/styling.ts`
