+++ packages/core/test/render3/perf/map_based_style_and_class_bindings/index.ts
@@ -8,9 +8,9 @@
 import {ɵɵadvance} from '../../../../src/render3/instructions/advance';
 import {ɵɵelement, ɵɵelementEnd, ɵɵelementStart} from '../../../../src/render3/instructions/element';
 import {refreshView} from '../../../../src/render3/instructions/shared';
+import {ɵɵclassMap, ɵɵstyleMap} from '../../../../src/render3/instructions/styling';
 import {RenderFlags} from '../../../../src/render3/interfaces/definition';
 import {TVIEW} from '../../../../src/render3/interfaces/view';
-import {ɵɵclassMap, ɵɵstyleMap} from '../../../../src/render3/styling/instructions';
 import {createBenchmark} from '../micro_bench';
 import {setupRootViewWithEmbeddedViews} from '../setup';
